+++ spec/language/fixtures/method_reference.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module MethodReference
+  def self.positive?(num)
+    num > 0
+  end
+
+  def self.add(a, b)
+    a + b
+  end
+end